Moved tests for webcasts to Webcasts folder
