+++ NetDevPL.Apps.WebApp.Tests/Features/Webcasts/WebcastsViewModelTests.cs
@@ -3,7 +3,7 @@ using NetDevPLWeb.Features.WebCasts;
 using Ploeh.AutoFixture;
 using Xunit;
 
-namespace NetDevPL.Apps.WebApp.Tests.Features
+namespace NetDevPL.Apps.WebApp.Tests.Features.Webcasts
 {
     public class WebcastsViewModelTests
     {
